requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

ipython == 7.12.0
ipywidgets == 7.5.1
matplotlib == 3.5.1
numpy == 1.19.5
pygame == 2.1.2
Pillow == 7.0.0
